# dreamtale-fast/management-web/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: dreamtale-fast是一个轻量级的Java快速开发平台，能快速开发项目并交付【接私活利器】
  version: 1.0.0
  title: 人人快速开发平台
basePath: /dreamtale-fast
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/user/list:
    get:
      tags:
        - 管理员管理
      summary: 管理员列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 用户名
          in: query
          type: string
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 管理员管理
      summary: 获取管理员信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回管理员信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 管理员管理
      summary: 修改密码
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 管理员管理
      summary: 添加管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 管理员管理
      summary: 修改管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 管理员管理
      summary: 删除管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密码
        type: string
      newPassword:
        description: 新密码
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
      type: object
      properties:
        userId:
          description: 用户ID
          type: integer
          format: int64
        username:
          description: 用户名
          type: string
        password:
          description: 密码
          type: string
        email:
          description: 邮箱
          type: string
        mobile:
          description: 手机号
          type: string
        status:
          description: 状态  0：禁用   1：正常
          type: integer
          format: int32
        roleIdList:
          description: 角色ID列表
          type: array
          items:
            type: integer
            format: int32
...